Refactor LabeledInput into separate files
