+++ src/components/RefundBankAccount.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from 'react'
 import { useForm } from 'react-hook-form'
 import { Box } from '@material-ui/core'
 import { useDispatch, useSelector } from 'react-redux'
-import { LabeledInput } from './labeledInput'
+import { LabeledInput } from './input/LabeledInput'
 import { Patterns } from './Patterns'
 import { saveRefundInfo } from '../redux/actions'
