+++ src/components/W2EmployerInfo.tsx
@@ -4,7 +4,7 @@ import { Box, Grow } from '@material-ui/core'
 import { useDispatch, useSelector } from 'react-redux'
 import countries from '../data/countries'
 
-import { LabeledInput, LabeledDropdown, LabeledCheckBox, USStateDropDown } from './labeledInput'
+import { LabeledInput, LabeledDropdown, LabeledCheckBox, USStateDropDown } from './input'
 import { Patterns } from './Patterns'
 import { saveEmployerData } from '../redux/actions'
 import { PagedFormProps } from './pager'
